requirements.txt
flask==3.0.2
openai-whisper==20231117
werkzeug==3.0.1
gunicorn==21.2.0
torch>=2.0.0
numpy>=1.20.0
ffmpeg-python>=0.2.0 